refactor to more generic implementation
